Turned operations into functions.
Functions and operations shouldn't have been two different things: Technically, operations are functions. By adopting the proper definition of operation, the design and the low-level implementation of the so-far-called "developer functions" and "operations" can be merged into the notion of "functions".

The code became much simpler thanks to this change. And things make more sense now.

Note that some tests break because of this.